+++ python/lib/candidate.py
@@ -16,7 +16,7 @@ class Candidate:
     :Example:
 
         from lib.candidate import Candidate
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
