+++ python/lib/database_lib/config.py
@@ -11,7 +11,7 @@ class Config:
     :Example:
 
         from lib.database_lib.config import Config
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
