+++ python/lib/database_lib/files.py
@@ -12,7 +12,7 @@ class Files:
     :Example:
 
         from lib.files import Files
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
