+++ python/lib/database_lib/mri_candidate_errors.py
@@ -11,7 +11,7 @@ class MriCandidateErrors:
     :Example:
 
         from lib.mri_candidate_errors import MriCandidateErrors
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
