+++ python/lib/database_lib/mri_protocol_checks.py
@@ -11,7 +11,7 @@ class MriProtocolChecks:
     :Example:
 
         from lib.mri_protocol_checks import MriProtocolChecks
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
