+++ python/lib/database_lib/mri_scan_type.py
@@ -11,7 +11,7 @@ class MriScanType:
     :Example:
 
         from lib.mri_scan_type import MriScanType
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
