+++ python/lib/database_lib/mri_scanner.py
@@ -13,7 +13,7 @@ class MriScanner:
     :Example:
 
         from lib.mri_scanner import MriScanner
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
