+++ python/lib/database_lib/notification.py
@@ -12,7 +12,7 @@ class Notification:
     :Example:
 
         from lib.notification import Notification
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
