+++ python/lib/database_lib/session_db.py
@@ -11,7 +11,7 @@ class SessionDB:
     :Example:
 
         from lib.database_lib.session_db import SessionDB
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
