+++ python/lib/database_lib/site.py
@@ -11,7 +11,7 @@ class Site:
     :Example:
 
         from lib.site import Site
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
