+++ python/lib/database_lib/tarchive_series.py
@@ -11,7 +11,7 @@ class TarchiveSeries:
     :Example:
 
         from lib.tarchive_series import TarchiveSeries
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
