+++ python/lib/database_lib/visit_windows.py
@@ -11,7 +11,7 @@ class VisitWindows:
     :Example:
 
         from lib.visit_windows import VisitWindows
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
