+++ python/lib/imaging.py
@@ -32,7 +32,7 @@ class Imaging:
     :Example:
 
         from lib.imaging  import Imaging
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
