+++ python/lib/imaging_upload.py
@@ -13,7 +13,7 @@ class ImagingUpload:
     :Example:
 
         from lib.imaging_upload  import ImagingUpload
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
