+++ python/lib/session.py
@@ -15,7 +15,7 @@ class Session:
     :Example:
 
         from lib.session import Session
-        from lib.database_mysqldb import Database
+        from lib.database_lib import Database
 
         # database connection
         db = Database(config.mysql, verbose)
