+++ python/mass_electrophysiology_chunking.py
@@ -6,7 +6,7 @@ import os
 import sys
 import getopt
 import lib.exitcode
-from lib.database_mysqldb import Database
+from lib.database_lib import Database
 from lib.physiological import Physiological
 from lib.database_lib.config import Config
