+++ tools/correct_lists_incorrectly_saved_in_parameter_file.py
@@ -2,7 +2,7 @@
 
 import os
 
-from lib.database_mysqldb import Database
+from lib.database_lib import Database
 from lib.lorisgetopt import LorisGetOpt
